+++ src/main/java/org/sonarsource/plugins/rtechnicaldebt/measures/cumulative/ComputeAverageDAM.java
@@ -25,14 +25,14 @@ public class ComputeAverageDAM implements MeasureComputer {
     // measure is already defined on files by {@link SetSizeOnFilesSensor}
     // in scanner stack
     if (context.getComponent().getType() != Component.Type.FILE) {
-      float sum = 0;
+      double sum = 0;
       int count = 0;
       for (Measure child : context.getChildrenMeasures(DATA_ACCESS_METRICS.key())) {
-        sum += child.getIntValue();
+        sum += child.getDoubleValue();
         count++;
       }
-      float average = count == 0 ? 0 : sum / count;
-      context.addMeasure(DATA_ACCESS_METRICS.key(), average);
+      double average = count == 0 ? 0 : sum / count;
+      context.addMeasure(DATA_ACCESS_METRICS.key(), (float) average);
     }
   }
 }
